refactor(core): move service setup to utils/
